Fix output being swallowed
`perform-build-step` was only capturing the output of the last command
in the closure, which was not what I wanted. Plus, we're unable to
capture the output of built-in commands this way, and providing the
command to run as a string makes it impossible to do that accidentally.

To make the non-captured output easier to see, captured output is no
longer printed to the console.